chore: move the snapshot description to shared package
