+++ x-pack/platform/packages/shared/kbn-data-lifecycle-phases/index.ts
@@ -15,3 +15,4 @@ export * from './src/enterprise_gating_modal';
 export * from './src/default_snapshot_repository_required_modal';
 export * from './src/edit_failed_data_lifecycle_flyout';
 export * from './src/phase_requirement_badges';
+export * from './src/searchable_snapshot_repository_info';
