+++ x-pack/platform/packages/shared/kbn-data-lifecycle-phases/src/searchable_snapshot_repository_info/index.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { SearchableSnapshotRepositoryInfo } from './searchable_snapshot_repository_info';
