+++ x-pack/platform/packages/shared/kbn-data-lifecycle-phases/tsconfig.json
@@ -20,6 +20,7 @@
   ],
   "kbn_references": [
     "@kbn/i18n",
+    "@kbn/i18n-react",
     "@kbn/storybook",
     "@kbn/index-lifecycle-management-common-shared",
     "@kbn/test-jest-helpers"
